comlete

+++ apps/todo/views.py
@@ -4,26 +4,30 @@ from django.forms import model_to_dict
 from .models import Todo,Task
 from django.contrib.auth.models import User
 import json
-
+from base64 import b64decode
+from django.contrib.auth import authenticate
 class TodosView(View):
-    def get(self, request: HttpRequest, user_id: int) -> HttpRequest:
-        try:
-            user = User.objects.get(username=user_id)
-        except User.DoesNotExist:
-            return JsonResponse({'error': 'user not found.'})
-        todos = Todo.objects.filter(user=user)
-
+    def get(self, request: HttpRequest) -> HttpRequest:
+        header=request.headers
+        auth=header["Authorization"][6:]
+        username, password=b64decode(auth).decode().split(":")
+        user=authenticate(username=username,password=password)
+        if user is None:
+            return JsonResponse({"error":"unauthorized."},status=401)
+        todos=Todo.objects.filter(user=user)
         result = []
         for todo in todos:
             result.append(model_to_dict(todo))
         
         return JsonResponse(result, safe=False)
-    
-    def post(self, request: HttpRequest, user_id: int):
-        try:
-            user = User.objects.get(id=user_id)
-        except User.DoesNotExist:
-            return JsonResponse({'error': 'user not found.'})
+     
+    def post(self, request: HttpRequest):
+        header=request.headers
+        auth=header["Authorization"][6:]
+        username, password=b64decode(auth).decode().split(":")
+        user=authenticate(username=username,password=password)
+        if user is None:
+            return JsonResponse({"error":"unauthorized."},status=401)
         
         data = json.loads(request.body.decode())
 
@@ -33,13 +37,15 @@ class TodosView(View):
         )
 
         return JsonResponse({"message": "Created"})
-
-    def put(self, request: HttpRequest, user_id: int, todo_id: int):
-        try:
-            user = User.objects.get(id=user_id)
-            todo = Todo.objects.get(id=todo_id, user=user)
-        except (User.DoesNotExist, Todo.DoesNotExist):
-            return JsonResponse({'error': 'user or todo not found.'})
+class TodoDetailsview(View):
+    def put(self, request: HttpRequest, todo_id: int):
+        header=request.headers
+        auth=header["Authorization"][6:]
+        username, password=b64decode(auth).decode().split(":")
+        user=authenticate(username=username,password=password)
+        if user is None:
+            return JsonResponse({"error":"unauthorized."},status=401)
+        todo=Todo.objects.get(user=user,id=todo_id)
         
         data = json.loads(request.body.decode())
         todo.title = data.get('title', todo.title)
@@ -47,26 +53,31 @@ class TodosView(View):
 
         return JsonResponse({"message": "Updated"})
 
-    def delete(self, request: HttpRequest, user_id: int, todo_id: int):
-        try:
-            user = User.objects.get(id=user_id)
-            todo = Todo.objects.get(id=todo_id, user=user)
-        except (User.DoesNotExist, Todo.DoesNotExist):
-            return JsonResponse({'error': 'user or todo not found.'})
-        
+    def delete(self, request: HttpRequest, todo_id: int):
+        header=request.headers
+        auth=header["Authorization"][6:]
+        username, password=b64decode(auth).decode().split(":")
+        user=authenticate(username=username,password=password)
+        if user is None:
+            return JsonResponse({"error":"unauthorized."},status=401)
+        todo=Todo.objects.get(user=user,id=todo_id)
         todo.delete()
 
         return JsonResponse({"message": "Deleted"})
 
 
 class TasksView(View):
-    def get(self, request: HttpRequest, user_id: int, todo_id: int):
+    def get(self, request: HttpRequest,  todo_id: int):
+        header=request.headers
+        auth=header["Authorization"][6:]
+        username, password=b64decode(auth).decode().split(":")
+        user=authenticate(username=username,password=password)
+        if user is None:
+            return JsonResponse({"error":"unauthorized."},status=401)
         try:
-            user = User.objects.get(id=user_id)
-            todo = Todo.objects.get(id=todo_id, user=user)
-        except (User.DoesNotExist, Todo.DoesNotExist):
-            return JsonResponse({'error': 'user or todo not found.'})
-
+            todo=Todo.objects.get(user=user,id=todo_id)
+        except Todo.DoesNotExist:
+            return JsonResponse({"error":"Todo does not exist"})
         tasks = Task.objects.filter(todo=todo)
         result = []
         for task in tasks:
@@ -74,46 +85,60 @@ class TasksView(View):
 
         return JsonResponse(result, safe=False)
 
-    def post(self, request: HttpRequest, user_id: int, todo_id: int):
+    def post(self, request: HttpRequest, todo_id: int):
+        header=request.headers
+        auth=header["Authorization"][6:]
+        username, password=b64decode(auth).decode().split(":")
+        user=authenticate(username=username,password=password)
+        if user is None:
+            return JsonResponse({"error":"unauthorized."},status=401)
         try:
-            user = User.objects.get(id=user_id)
-            todo = Todo.objects.get(id=todo_id, user=user)
-        except (User.DoesNotExist, Todo.DoesNotExist):
-            return JsonResponse({'error': 'user or todo not found.'})
-
+            todo=Todo.objects.get(user=user,id=todo_id)
+        except Todo.DoesNotExist:
+            return JsonResponse({"error":"Todo does not exist"})
         data = json.loads(request.body.decode())
 
         Task.objects.create(
             title=data.get('title'),
-            completed=data.get('completed'),
+            description=data.get("description"),
             todo=todo,
         )
 
         return JsonResponse({"message": "Created"})
-
-    def put(self, request: HttpRequest, user_id: int, todo_id: int, task_id: int):
+class TaskDetailsView(View):
+    def put(self, request: HttpRequest, todo_id: int, task_id: int):
+        header=request.headers
+        auth=header["Authorization"][6:]
+        username, password=b64decode(auth).decode().split(":")
+        user=authenticate(username=username,password=password)
+        if user is None:
+            return JsonResponse({"error":"unauthorized."},status=401)
         try:
-            user = User.objects.get(id=user_id)
-            todo = Todo.objects.get(id=todo_id, user=user)
-            task = Task.objects.get(id=task_id, todo=todo)
-        except (User.DoesNotExist, Todo.DoesNotExist, Task.DoesNotExist):
-            return JsonResponse({'error': 'user, todo, or task not found.'})
+            todo=Todo.objects.get(user=user,id=todo_id)
+            task=Task.objects.get(id=task_id,todo=todo)
+        except (Todo.DoesNotExist, Task.DoesNotExist):
+            return JsonResponse({"error":"Todo does not exist or task does not exist"})
 
         data = json.loads(request.body.decode())
         task.title = data.get('title', task.title)
-        task.completed = data.get('completed', task.completed)
+        task.description=data.get("Description",task.description)
         task.save()
 
         return JsonResponse({"message": "Updated"})
 
-    def delete(self, request: HttpRequest, user_id: int, todo_id: int, task_id: int):
+    def delete(self, request: HttpRequest, todo_id: int, task_id: int):
+        header=request.headers
+        auth=header["Authorization"][6:]
+        username, password=b64decode(auth).decode().split(":")
+        user=authenticate(username=username,password=password)
+        if user is None:
+            return JsonResponse({"error":"unauthorized."},status=401)
         try:
-            user = User.objects.get(id=user_id)
-            todo = Todo.objects.get(id=todo_id, user=user)
-            task = Task.objects.get(id=task_id, todo=todo)
-        except (User.DoesNotExist, Todo.DoesNotExist, Task.DoesNotExist):
-            return JsonResponse({'error': 'user, todo, or task not found.'})
-        
+            todo=Todo.objects.get(user=user,id=todo_id)
+            task=Task.objects.get(id=task_id,todo=todo)
+        except (Todo.DoesNotExist, Task.DoesNotExist):
+            return JsonResponse({"error":"Todo does not exist or task does not exist"})
         task.delete()
 
         return JsonResponse({"message": "Deleted"})
+
